test: resolve Match.test.ts collision onto the comprehensive suite
The legacy Match suite (~35 cases) and a co-located 2-case suite (from the
draw-score fix 10551e3c) both tested the same Match source; the 2-case suite
was a strict subset. Keep the comprehensive suite co-located in src/, drop the
2-case file, and fold its only extra assertions (isFinished()===false after a
draw and after a win) into the matching cases. Fix a misleading title:
'should increment both scores when draw occurs' already asserted 0-0, renamed
to 'should NOT increment scores when a draw occurs'. No coverage lost.
